Refactor: timetable component handleSelect함수 분리
